Add files via upload
It is an advanced scientific calculator developed with python programming language using python packages(built-in and external): math, statistics, numpy, scipy, sympy and functools. With the aid of this calculator problems on linear system of equations, roots of polynomial functions, determinant of matrices with order higher than 3 can be solved using this calculator. Limits, derivatives, and integration can easily be handled with the aid of the calculator and problems in statistics. and many more.